# compress: add a deflate compressor