@squeek502 2022-01-10
I assume you're going to need a *lot* more reasoning (benchmarks? use-cases?) for wholesale replacing the previous inflate implementation.