@hdorio 2022-01-10
> you're going to need a _lot_ more reasoning

Not really, any [puff.c](https://github.com/madler/zlib/blob/master/contrib/puff/puff.c) based implementation should be replaced.

    puff.c is a simple inflate written to be an unambiguous way to specify the
    deflate format.  It is not written for speed but rather simplicity.

> benchmarks?

You mean like this one? https://github.com/ziglang/zig/commit/41f244bd2ff0a217a83a3d17fc82f4a773eca278
They are not really useful in my opinion but if I must: https://gist.github.com/hdorio/5e4f26150198a7515d89f01be5e0a45f

```
// outputs
//
// deflate_old: 3.982s
// deflate_new: 2.08s
```

The Go implementation is based off https://github.com/madler/zlib/raw/master/doc/algorithm.txt .

I invite you to do your own benchmarks with your own sample data, just keep in mind this new implementation does a lot of small reads for the decompression, so don't pass in a file writer directly, use a buffer or read the whole input into memory to avoid lots of small fread() system calls.