@squeek502 2022-01-11
I'm not able to run your benchmark with the latest commit of this branch, did something in the compressor change?

(note: I'm running with ReleaseSafe just to get a stacktrace):

```
$ zig run bench.zig -O ReleaseSafe
error: InvalidStoredSize
/home/ryan/Programming/zig/zig/lib/std/os.zig:0:5: 0x2176ee in main (bench)
/home/ryan/Programming/zig/zig/lib/std/fs.zig:2055:9: 0x2176f6 in main (bench)
        return result;
        ^
/home/ryan/Programming/zig/zig/lib/std/fs.zig:2036:9: 0x2176fe in main (bench)
        return self.accessZ(&path_c, flags);
        ^
/home/ryan/Programming/zig/zig/lib/std/compress/deflate/decompressor.zig:0:37: 0x21a514 in main (bench)
/home/ryan/Programming/zig/zig/lib/std/compress/deflate_old.zig:0:34: 0x218ed3 in main (bench)
/home/ryan/Programming/zig/zig/lib/std/compress/deflate_old.zig:668:17: 0x218edb in main (bench)
                try self.step();
                ^
/home/ryan/Programming/zig/zig/lib/std/io/reader.zig:26:13: 0x218ee3 in main (bench)
            return readFn(self.context, buffer);
            ^
/home/ryan/Programming/zig/zig/lib/std/os/linux.zig:0:17: 0x219135 in main (bench)
/home/ryan/Programming/zig/zig/lib/std/Thread/Mutex.zig:0:35: 0x21a415 in main (bench)
```

and with the old deflate commented out, I still get an error when decompressing:

```
$ zig run bench.zig -O ReleaseSafe
error: CorruptInput
/home/ryan/Programming/zig/zig/lib/std/compress/deflate/decompressor.zig:0:0: 0x224d2f in std.compress.deflate.decompressor.Decompressor(std.io.reader.Reader(*std.io.fixed_buffer_stream.FixedBufferStream([]u8),std.io.fixed_buffer_stream.ReadError,std.io.fixed_buffer_stream.FixedBufferStream([]u8).read)).nextBlock (bench)
/home/ryan/Programming/zig/zig/lib/std/compress/deflate/decompressor.zig:0:0: 0x224d37 in std.compress.deflate.decompressor.Decompressor(std.io.reader.Reader(*std.io.fixed_buffer_stream.FixedBufferStream([]u8),std.io.fixed_buffer_stream.ReadError,std.io.fixed_buffer_stream.FixedBufferStream([]u8).read)).nextBlock (bench)
/home/ryan/Programming/zig/zig/lib/std/compress/deflate/decompressor.zig:0:38: 0x2082a7 in std.start.posixCallMainAndExit (bench)
/home/ryan/Programming/zig/zig/lib/std/io/reader.zig:26:13: 0x2087e7 in std.start.posixCallMainAndExit (bench)
            return readFn(self.context, buffer);
            ^
/home/ryan/Programming/zig/tmp/deflate-bench/bench.zig:0:0: 0x206808 in std.start.posixCallMainAndExit (bench)
/home/ryan/Programming/zig/tmp/deflate-bench/bench.zig:0:0: 0x206481 in std.start.posixCallMainAndExit (bench)
```