@hdorio 2022-01-12
@andrew thanks for taking some time on this

At first, your idea seemed good to me but after some preliminary tests, I would advise against putting the slow implementation under -OReleaseSmall.
because
- The gain in size is not clear.
- The performance gap is too big. As a user I don't expect ReleaseSmall to be so aggressive about memory usage for such poor performance.
- ReleaseSmall does not say anything about fixed/dynamic heap memory allocation or how the memory is used. (currently only says: (Medium runtime performance, Safety checks disabled, Slow compilation speed, Small binary size, Reproducible build)

I think if we want to keep the slow one, we should let the user choose between the two.
Or optimize further the slow one to have a real gain in binary size.
Or add what ReleaseSmall is about, about memory allocation.

**The gain in size is not clear**

When both implementations use an allocator (for reading the whole file in memory) the code size of the fast implementation is smaller of ~3k (22984), when an allocator is not used for the slow one, the slow one is smaller of 2k (20144), when a custom allocator is used for the fast one, the fast one is smaller of 3k (17032).
So it seems the code for how the allocator is implemented is the problem.

<details>
  <summary>binaries sizes (-OReleaseSmall)</summary>

```
zig build-exe -OReleaseSmall

.strip = strip command was used
.zigstrip = zig build-exe --strip
.custom-alloc = use a fixed buffer allocator of 3mb from a stack buffer
.no-alloc = don't preload the file into memory
.buffered-reader = use the buffered-reader from std to read the file
.new = fast implementation
.old = slow implementation

-rwxr-xr-x  252584 decompress.new
-rwxr-xr-x  238312 decompress.old
-rwxr-xr-x   22984 decompress.strip.new
-rwxr-xr-x   26640 decompress.strip.old
-rwxr-xr-x   30568 decompress.zigstrip.new
-rwxr-xr-x   32848 decompress.zigstrip.old
-rwxr-xr-x   17032 decompress.zigstrip.strip.custom-alloc.new
-rwxr-xr-x   22712 decompress.zigstrip.strip.new
-rwxr-xr-x  128480 decompress.zigstrip.strip.no-alloc.buffered-reader.old
-rwxr-xr-x   20144 decompress.zigstrip.strip.no-alloc.old
-rwxr-xr-x   26240 decompress.zigstrip.strip.old
```
</details>

**The performance gap**

_decompress.zigstrip.strip.no-alloc.buffered-reader.old vs decompress.zigstrip.strip.custom-alloc.new_

I'm sure the slow one is perfect for some environments where memory constraints are very harsh. 348 kbytes (slow) vs 3120 kbytes (fast) that's very good but the time for a 660mb file of 1m2.166s (slow) vs 23.012s (new) is not going from good to medium runtime performance to me but from ok to poor (it's subjective). Here's the thing... the fast implementation is not that good. I expect someone to do much better than this, like around 2x the current inflate speed. And if the zlib documentation is correct, I expect someone to do this for a constant amount of memory of 40kb instead of around 3mb for mine.

<details>
  <summary>Maximum resident set size (time -v)</summary>

```
calgary$ /usr/bin/time -v ./decompress.zigstrip.strip.old
deflate_old: 9.447s
	Command being timed: "./decompress.zigstrip.strip.old"
	User time (seconds): 6.86
	System time (seconds): 1.14
	Percent of CPU this job got: 84%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:09.44
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 1106112
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 0
	Minor (reclaiming a frame) page faults: 578963
	Voluntary context switches: 59
	Involuntary context switches: 865
	Swaps: 0
	File system inputs: 0
	File system outputs: 1360568
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 0
	Page size (bytes): 4096
	Exit status: 0
calgary$ /usr/bin/time -v ./decompress.zigstrip.strip.new
deflate_new: 5.531s
	Command being timed: "./decompress.zigstrip.strip.new"
	User time (seconds): 2.60
	System time (seconds): 1.43
	Percent of CPU this job got: 72%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:05.53
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 1106104
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 0
	Minor (reclaiming a frame) page faults: 598491
	Voluntary context switches: 75
	Involuntary context switches: 417
	Swaps: 0
	File system inputs: 0
	File system outputs: 1360568
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 0
	Page size (bytes): 4096
	Exit status: 0
	
	
calgary$ /usr/bin/time -v ./decompress.zigstrip.strip.no-alloc.old 
deflate_old: 4m446.975ms
	Command being timed: "./decompress.zigstrip.strip.no-alloc.old"
	User time (seconds): 114.90
	System time (seconds): 125.30
	Percent of CPU this job got: 99%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 4:00.44
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 352
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 0
	Minor (reclaiming a frame) page faults: 38
	Voluntary context switches: 78
	Involuntary context switches: 23703
	Swaps: 0
	File system inputs: 0
	File system outputs: 1360568
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 0
	Page size (bytes): 4096
	Exit status: 0
	
calgary$ /usr/bin/time -v ./decompress.zigstrip.strip.no-alloc.buffered-reader.old
deflate_old: 1m2.166s
	Command being timed: "./decompress.zigstrip.strip.no-alloc.buffered-reader.old"
	User time (seconds): 60.80
	System time (seconds): 0.67
	Percent of CPU this job got: 98%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 1:02.16
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 348
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 0
	Minor (reclaiming a frame) page faults: 42
	Voluntary context switches: 50
	Involuntary context switches: 6079
	Swaps: 0
	File system inputs: 0
	File system outputs: 1360568
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 0
	Page size (bytes): 4096
	Exit status: 0
	
calgary$ /usr/bin/time -v ./decompress.zigstrip.strip.custom-alloc.new
deflate_new: 23.012s
	Command being timed: "./decompress.zigstrip.strip.custom-alloc.new"
	User time (seconds): 7.77
	System time (seconds): 15.00
	Percent of CPU this job got: 98%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:23.01
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 3120
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 0
	Minor (reclaiming a frame) page faults: 785
	Voluntary context switches: 47
	Involuntary context switches: 2272
	Swaps: 0
	File system inputs: 0
	File system outputs: 1360568
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 0
	Page size (bytes): 4096
	Exit status: 0
```

</details>

**unifying the two interfaces**

> @hdorio what do you think? Can the ported go code be adjusted to require a constant amount of memory?

Last time I tried, I failed. But It seems it's technically possible, zlib uses the same algorithm and end up with this
https://zlib.net/zlib_tech.html
```
inflate memory usage (bytes) = (1 << windowBits) + 7 KB
Typically, therefore, inflate() requires no more than 40 KB of storage on a 64-bit machine. This includes the 32768-byte sliding window and approximately 7 KB for the inflated data structure.
```

> Or perhaps could the scratch memory buffer be provided by the caller for both implementations, thereby unifying the two interfaces?

I don't know how much memory would be needed for the scratch memory buffer, this depends on the stream being decompressed.

Also the slow one would need to be modified to be able to take a dictionary.

**bikeshedding**

> - simple
> - fast

I also prefer `simple / fast` but what is considered fast today may be considered slow tomorrow. 
As I said, I expect someone to do much better than the one in this pull request.

> - compress/deflate/simple/inflate_stream.zig
> - compress/deflate/fast/inflate_stream.zig

The whole shrink/unshrink, implode/explode, deflate/inflate trend is cute but as a non-native english speaker I would much prefer to have to remember the name of the algorithm (deflate) then choose to decompress_stream or compress_stream instead of remembering the opposite of 'deflate' is 'inflate'.